pyproject.toml
[project]
name = "haul"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.7",
    "fastmcp>=3.2.1",
    "httpx>=0.28.1",
    "kyber-py>=1.2.0",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
